+++ app/database.py
@@ -5,4 +5,3 @@ client = MongoClient(settings.MONGO_URI)
 db = client[settings.DB_NAME]
 
 users_collection = db["users"]
-organizations_collection = db["organizations"]
