+++ app/routes/notify.py
@@ -1,5 +1,5 @@
 from fastapi import APIRouter, Depends
-from app.database import users_collection, organizations_collection
+from app.database import users_collection
 from app.utils.emailer import send_email
 from app.utils.security import verify_api_key
 
@@ -8,9 +8,7 @@ router = APIRouter()
 @router.post("/launch")
 def notify_all(_: bool = Depends(verify_api_key)):
     users = list(users_collection.find({}, {"_id": 0, "email": 1}))
-    orgs = list(organizations_collection.find({}, {"_id": 0, "email": 1}))
-
-    recipients = [u["email"] for u in users] + [o["email"] for o in orgs]
+    recipients = [u["email"] for u in users]
 
     for email in recipients:
         send_email(
@@ -18,4 +16,4 @@ def notify_all(_: bool = Depends(verify_api_key)):
             subject="Valido is Live! 🚀",
             body="We’re excited to announce that Valido is live! Sign up now at https://valido.com"
         )
-    return {"message": f"Notification sent to {len(recipients)} recipients"}
+    return {"message": f"Notification sent to {len(recipients)} users"}
